Reorganize tests for Netv4Addr and split impls into multiple files
