+++ src/netv4addr/from.rs
@@ -1,0 +1,2 @@
+use super::Netv4Addr;
+mod ipv4addr;
