+++ src/netv4addr/from/ipv4addr.rs
@@ -1,0 +1,8 @@
+use super::Netv4Addr;
+use std::net::Ipv4Addr;
+
+impl From<Ipv4Addr> for Netv4Addr {
+	fn from(addr: Ipv4Addr) -> Self {
+		Self::new(addr, Ipv4Addr::from(u32::max_value()))
+	}
+}
