+++ src/netv4addr/hash.rs
@@ -1,20 +1,19 @@
-use crate::*;
-
-mod netv6addr {
-	use super::Netv6Addr;
+#[cfg(test)]
+mod tests {
+	use crate::Netv4Addr;
+	use std::hash::Hash;
 
 	#[test]
 	fn hash_same() {
 		use std::collections::hash_map::DefaultHasher;
-		use std::hash::Hash;
 		let mut hasher = DefaultHasher::new();
 		assert_eq!(
-			"2001:db8:dead:beef::1325/64"
-				.parse::<Netv6Addr>()
+			"192.0.2.26/29"
+				.parse::<Netv4Addr>()
 				.unwrap()
 				.hash(&mut hasher),
-			"2001:db8:dead:beef::1325/64"
-				.parse::<Netv6Addr>()
+			"192.0.2.26/29"
+				.parse::<Netv4Addr>()
 				.unwrap()
 				.hash(&mut hasher)
 		);
