+++ src/netv4addr/partialord.rs
@@ -1,0 +1,12 @@
+use super::Netv4Addr;
+use core::cmp::Ordering;
+
+impl PartialOrd for Netv4Addr {
+	fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+		match self.addr().partial_cmp(other.addr()) {
+			Some(Ordering::Equal) => self.mask().partial_cmp(other.mask()),
+			Some(ordering) => Some(ordering),
+			None => None,
+		}
+	}
+}
